+++ upkeep/utils/__init__.py
@@ -2,5 +2,6 @@
 from __future__ import annotations
 
 from .misc import CommandRunner, minenv
+from .portage import binary_package_directory, build_directory, portage_env
 
-__all__ = ('CommandRunner', 'minenv')
+__all__ = ('CommandRunner', 'binary_package_directory', 'build_directory', 'minenv', 'portage_env')